GroupProject/requirements.txt
kafka-python==2.0.2
pandas==2.0.3
openpyxl==3.1.2
langdetect==1.0.9
textblob==0.17.1
nltk==3.8.1
pyarrow==15.0.0